pyproject.toml
[tool.poetry]
name = "ENCOT"
version = "1.0.0"
description = "Transformer-based codon optimization for E. coli using deep learning with Augmented-Lagrangian GC control."
authors = ["Adibvafa Fallahpour <Adibvafa.fallahpour@mail.utoronto.ca>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/geno543/ENCOT"
repository = "https://github.com/geno543/ENCOT"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

[tool.poetry.dependencies]
python = "^3.9"
biopython = "^1.83"
ipywidgets = "^7.0.0"
numpy = "<2.0.0"
onnxruntime = "^1.16.3"
pandas = "^2.0.0"
python_codon_tables = "^0.1.12"
pytorch_lightning = "^2.2.1"
scikit-learn = "^1.2.2"
scipy = "^1.13.1"
setuptools = "^70.0.0"
torch = "^2.0.0"
tqdm = "^4.66.2"
transformers = "^4.40.0"
CAI-PyPI = "^2.0.1"
codon-bias = "^1.0.2"
gcua = "^0.1.2"
dtw-python = "^1.3.0"

[tool.poetry.dev-dependencies]
coverage = {version = "^7.0", extras = ["toml"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.coverage.run]
omit = [
    # omit pytorch-generated files in /tmp
    "/tmp/*",
]
